==> 04_phenotypes/01_ICD10_extraction.R <==
# Extract phenotype information for binary disease outcomes and generate a PLINK-compatible file

#################################################
### Libraries ###################################
library(data.table)
library(dplyr)
library(tidyr)
library(stringr)
library(tibble)



#################################################
### Load Main Data ##############################

# Phenotype list with inclusion and exclusion criteria for each disease; Can be generated from Table S1
pheno_list <- as.data.frame(fread("/data/phenotypes/ICD10_binary_traits_definition.txt", header = T))



#################################################
### STEP 1: Select relevant phenotype columns ###

# Self-reported data: cancer (SRC) and other diseases (SRD); "ukb44073.csv" contains phenotypic data for field 20001 (self-reported cancer; SRC) and 20002 (self-reported disease; SRD) available from the UKBB portal (downloaded: 10/2020)
header <- as.data.frame(fread("/data/UKBB/pheno/ukb44073.csv", header = T, nrow = 0))
col_SRC <- grep("^20001-", names(header))
col_SRD <- grep("^20002-", names(header))
rm(header)

# Hospital diagnosis data; "ukb52238.csv" contains phenotypic data for field 41270 (all ICD-10 codes) available from the UKBB portal (downloaded: 05/2022)
header <- as.data.frame(fread("/data/UKBB/pheno/ukb52238.csv", header = T, nrow = 0))
col_ICD <- grep("^41270-", names(header))
rm(header)



#################################################
### STEP 2: Generate index tables ###############
# That is a table with all the disease codes received by each individual in a long format; one per UKBB field 

# Self-reported cancer (#20001)
SRC <- as.data.frame(fread("/data/UKBB/pheno/ukb44073.csv", header = T, select = c(1, col_SRC), colClass = "character"))
SRC[SRC==""] <- NA
SRC <- na.omit(gather(SRC, Instance, Code, 2:ncol(SRC)))
SRC <- unique(SRC[, c(1,3)]) # 49'904 self-reported cancers

# Self-reported disease (#20002)
SRD <- as.data.frame(fread("/data/UKBB/pheno/ukb44073.csv", header = T, select = c(1, col_SRD), colClass = "character"))
SRD[SRD==""] <- NA
SRD <- na.omit(gather(SRD, Instance, Code, 2:ncol(SRD)))
SRD <- unique(SRD[, c(1,3)]) # 1'069'671 self-reported diseases

# ICD10 diagnosis (#41270)
ICD <- as.data.frame(fread("/data/UKBB/pheno/ukb52238.csv", header = T, select = c(1, col_ICD), colClass = "character"))
ICD[ICD==""] <- NA
ICD <- na.omit(gather(ICD, Instance, Code, 2:ncol(ICD)))
ICD <- unique(ICD[, c(1,3)]) # 6'302'095 ICD10 diagnosis 



#################################################
### STEP 3: Generate phenotype tables ###########
# That is a table with all the disease diagnoses as per our definition, combining information from the different fields (samples x diseases)

# Load white-british individuals; Generated by "01_samples/sample_filtering.R"
eid_all <- as.data.frame(fread("/cauwerx/gwas/01_samples/data/samples_white_british_All.txt", header = F, col.names = "eid"))

# Generate an empty table, where every individual is a control (i.e., 1) 
pheno_all <- as.data.frame(matrix(data = 1, nrow = nrow(eid_all), ncol = nrow(pheno_list), dimnames = list(c(1:nrow(eid_all)), c(pheno_list$phenotype))))
pheno_all <- cbind(eid_all, pheno_all)


### 3.1 Exclusion list ###########################
# Set individuals with a code/diagnosis on the exclusion list to missing (i.e., NA)

# Loop over diseases
for (p in 1:nrow(pheno_list)) {

	# Define the disease
	pheno <- pheno_list[p, "phenotype"]
	print(paste0("Starting to exclude individuals for ", pheno))

	# Excluded SRC (#20001)
	excl_SRC <- str_split(pheno_list[p, "exclude_20001"], ", ")[[1]]
	excl_SRC_eid <- unique(SRC[SRC$Code %in% excl_SRC, "eid"])

	# Excluded SRD (#20002)
	excl_SRD <- str_split(pheno_list[p, "exclude_20002"], ", ")[[1]]
	excl_SRD_eid <- unique(SRD[SRD$Code %in% excl_SRD, "eid"])

	# Excluded ICD (#41270)
	excl_ICD <- str_split(pheno_list[p, "exclude_41270"], ", ")[[1]]
	excl_ICD <- paste(paste0(sub("\\.", "", excl_ICD), ".*"), collapse = "|")
	excl_ICD_eid <- ICD %>% filter(str_detect(Code, excl_ICD))
	excl_ICD_eid <- unique(excl_ICD_eid$eid)

	# Merge excluded samples for different data fields
	excl <- as.numeric(unique(c(excl_SRC_eid, excl_SRD_eid, excl_ICD_eid)))
	print(paste0("Number of individuals excluded for ", pheno, ": ", length(excl)))
	
	# Correct the table by setting selected individuals as missing
	pheno_all[pheno_all$eid %in% excl, pheno] <- NA

}
rm()


### 3.2 Inclusion list ##########################
# Set individuals with a code on the inclusion list to cases (i.e., 2)

# Loop over diseases
for (p in 1:nrow(pheno_list)) {

	# Define disease
	pheno <- pheno_list[p, "phenotype"]
	print(paste0("Starting to include individuals for ", pheno))

	# Include ICD (#41270)
	incl_ICD <- str_split(pheno_list[p, "include_41270"], ", ")[[1]]
	incl_ICD <- paste(paste0(sub("\\.", "", incl_ICD), ".*"), collapse = "|")
	incl_ICD_eid <- ICD %>% filter(str_detect(Code, incl_ICD))
	incl_ICD_eid <- as.numeric(unique(incl_ICD_eid$eid))
	print(paste0("Number of individuals included for ", pheno, ": ", length(incl_ICD_eid)))
	
	# Correct the table by setting selected individuals as cases
	pheno_all[pheno_all$eid %in% incl_ICD_eid, pheno] <- 2

}



#################################################
### STEP 4: Calculate the disease burden ########

# Calculation is based on 55 non sex-specific diseases, with the disease burden being defined as the number of diagnoses received
exc_traits <- c("eid", "BC", "OC", "PC", "endometriosis", "menstruation")
pheno_all$DiseaseBurden <- rowSums(pheno_all[, !names(pheno_all) %in% c(exc_traits, "AnyDisease")] == 2, na.rm = T)
pheno_all <- pheno_all[, !names(pheno_all) %in% c("AnyDisease")]



#################################################
### STEP 5: Split by sex ########################

# Males (sex agnostic + male-specific diseases --> 56 diseases + burden)
eid_M <- as.data.frame(fread("/cauwerx/gwas/01_samples/data/samples_white_british_M.txt", header = F, col.names = "eid")) # File generated by "01_samples/sample_filtering.R"
pheno_M <- pheno_all[pheno_all$eid %in% eid_M$eid, !names(pheno_all) %in% c("BC", "OC", "endometriosis", "menstruation")]
colnames(pheno_M)[1] <- "IID"
print(paste0("Dimensions of disease table for unrelated white British males: ", ncol(pheno_M)-1, " pheno x ", nrow(pheno_M), " males"))

# Females (sex agnostic + female-specific diseases --> 59 diseases + burden)
eid_F <- as.data.frame(fread("/cauwerx/gwas/01_samples/data/samples_white_british_M.txt", header = F, col.names = "eid")) # File generated by "01_samples/sample_filtering.R"
pheno_F <- pheno_all[pheno_all$eid %in% eid_F$eid, !names(pheno_all) %in% c("PC")]
colnames(pheno_F)[1] <- "IID"
print(paste0("Dimensions of disease table for unrelated white British females: ", ncol(pheno_F)-1, " pheno x ", nrow(pheno_F), " females"))

# All (sex agnostic diseases --> 55 diseases + burden)
pheno_all <- pheno_all[ , !names(pheno_all) %in% c("BC", "OC", "PC", "endometriosis", "menstruation")]
colnames(pheno_all)[1] <- "IID"
print(paste0("Dimensions of disease table for unrelated white British individuals: ", ncol(pheno_all)-1, " pheno x ", nrow(pheno_all), " individuals"))



#################################################
### Save data ###################################

# For use with PLINK v2.0 - All
fwrite(pheno_all, "/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_All.txt", col.names = T, row.names = F, quote = F, sep = "\t", na = "NA")
# Males
fwrite(pheno_M, "/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_M.txt", col.names = T, row.names = F, quote = F, sep = "\t", na = "NA")
# Females
fwrite(pheno_F, "/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_F.txt", col.names = T, row.names = F, quote = F, sep = "\t", na = "NA")

# For use with PLINK v1.9 - All
pheno_all_plink_v1.9 <- add_column(pheno_all, FID = 0, .before = "IID")
fwrite(pheno_all_plink_v1.9, "/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_All_plink_v1.9.txt", col.names = T, row.names = F, quote = F, sep = "\t", na = "NA")
# Males
pheno_M_plink_v1.9 <- add_column(pheno_M, FID = 0, .before = "IID")
fwrite(pheno_M_plink_v1.9, "/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_M_plink_v1.9.txt", col.names = T, row.names = F, quote = F, sep = "\t", na = "NA")
# Females
pheno_F_plink_v1.9 <- add_column(pheno_F, FID = 0, .before = "IID")
fwrite(pheno_F_plink_v1.9, "/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_F_plink_v1.9.txt", col.names = T, row.names = F, quote = F, sep = "\t", na = "NA")
