==> 04_phenotypes/02_ICD10_prevalence.R <==
# Calculate the prevalence of each disease in a sex-combine dand sex-specific fashion 

#################################################
### Libraries ###################################
library(data.table)
library(dplyr)



#################################################
### Load data ###################################

# Phenotype data (by sex); Files generated by "04_phenotypes/01_ICD10_extraction.R"
pheno_M <- as.data.frame(fread("/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_M.txt", header = T))
pheno_F <- as.data.frame(fread("/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_F.txt", header = T))
pheno_All <- as.data.frame(fread("/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_All.txt", header = T))

# Create a union of phenotypes (--> 60 diseases + burden)
pheno_union <- c(names(pheno_M), names(pheno_F))
pheno_union <-  unique(pheno_union[grep("IID", pheno_union, invert = T)])



#################################################
### STEP 1: Prevalence & Fisher test ############

# Create an empty dataframe to save results from the prevalence calculation and Fisher tests (compare prevalence between sexes)
results_fisher <- data.frame()

# Counter
c <- 1

# Loop over diseases
for (p in pheno_union ) {

	print(paste0("Analyzing (prevalence & Fisher test): ", p))

	# Verify that the phenotype is available in both sexes and is not a composite phenotype
	if (!p %in% c("OC", "BC", "endometriosis", "menstruation", "PC", "DiseaseBurden")) {

		# Calculate the number of cases/controls for both sexes
		NumCtrl_M <- nrow(pheno_M[which(pheno_M[,p] == 1), ])
		NumCtrl_F <- nrow(pheno_F[which(pheno_F[,p] == 1), ])
		NumCase_M <- nrow(pheno_M[which(pheno_M[,p] == 2), ])
		NumCase_F <- nrow(pheno_F[which(pheno_F[,p] == 2), ])

		# Calculate the prevalence for both sexes
		prevalence_M <- NumCase_M/(NumCtrl_M + NumCase_M)
		prevalence_F <- NumCase_F/(NumCtrl_F + NumCase_F)
		prevalence_all <- (NumCase_M + NumCase_F)/(NumCase_M + NumCtrl_M + NumCase_F + NumCtrl_F)

		# Fisher tests comparing prevalence in males vs females
		fisher_OR  <- fisher.test(matrix(c(NumCase_M, NumCase_F, NumCtrl_M, NumCtrl_F), nrow = 2))$estimate
		fisher_CI_lower  <- fisher.test(matrix(c(NumCase_M, NumCase_F, NumCtrl_M, NumCtrl_F), nrow = 2))$conf.int[1]
		fisher_CI_upper  <- fisher.test(matrix(c(NumCase_M, NumCase_F, NumCtrl_M, NumCtrl_F), nrow = 2))$conf.int[2]
		fisher_p  <- fisher.test(matrix(c(NumCase_M, NumCase_F, NumCtrl_M, NumCtrl_F), nrow = 2))$p.value
		
		# Write results to table
		results_fisher[c, "PHENO"] <- p
		results_fisher[c, "PREV"] <- prevalence_all
		results_fisher[c, "PREV_M"] <- prevalence_M
		results_fisher[c, "PREV_F"] <- prevalence_F
		results_fisher[c, "CASE"] <- NumCase_M + NumCase_F
		results_fisher[c, "CASE_M"] <- NumCase_M
		results_fisher[c, "CASE_F"] <- NumCase_F
		results_fisher[c, "CTRL"] <- NumCtrl_M + NumCtrl_F
		results_fisher[c, "CTRL_M"] <- NumCtrl_M
		results_fisher[c, "CTRL_F"] <- NumCtrl_F
		results_fisher[c, "OR"] <- fisher_OR
		results_fisher[c, "LCI"] <- fisher_CI_lower
		results_fisher[c, "UCI"] <- fisher_CI_upper
		results_fisher[c, "P"] <- fisher_p

		# Increase counter
		c <- c + 1

	# If female-specific phenotype, only assess case number and prevalence in females
	} else if (p %in% c("OC", "BC", "endometriosis", "menstruation")) {

		# Calculate the number of cases/controls in females
		NumCtrl_F <- nrow(pheno_F[which(pheno_F[,p] == 1), ])
		NumCase_F <- nrow(pheno_F[which(pheno_F[,p] == 2), ])

		# Calculate the prevalence in females
		prevalence_F <- NumCase_F/(NumCtrl_F + NumCase_F)
		
		# Write results to table
		results_fisher[c, "PHENO"] <- p
		results_fisher[c, "PREV"] <- NA
		results_fisher[c, "PREV_M"] <- NA
		results_fisher[c, "PREV_F"] <- prevalence_F
		results_fisher[c, "CASE"] <- NA
		results_fisher[c, "CASE_M"] <- NA
		results_fisher[c, "CASE_F"] <- NumCase_F
		results_fisher[c, "CTRL"] <- NA
		results_fisher[c, "CTRL_M"] <- NA
		results_fisher[c, "CTRL_F"] <- NumCtrl_F
		results_fisher[c, "OR"] <- NA
		results_fisher[c, "LCI"] <- NA
		results_fisher[c, "UCI"] <- NA
		results_fisher[c, "P"] <- NA

		# Increase counter
		c <- c + 1

	# If male-specific phenotype, only assess case number and prevalence in males
	} else if (p %in% c("PC")) {

		# Calculate the number of cases/controls in males
		NumCtrl_M <- nrow(pheno_M[which(pheno_M[,p] == 1), ])
		NumCase_M <- nrow(pheno_M[which(pheno_M[,p] == 2), ])

		# Calculate the prevalence in males
		prevalence_M <- NumCase_M/(NumCtrl_M + NumCase_M)
		
		# Write results to table
		results_fisher[c, "PHENO"] <- p
		results_fisher[c, "PREV"] <- NA
		results_fisher[c, "PREV_M"] <- prevalence_M
		results_fisher[c, "PREV_F"] <- NA
		results_fisher[c, "CASE"] <- NA
		results_fisher[c, "CASE_M"] <- NumCase_M
		results_fisher[c, "CASE_F"] <- NA
		results_fisher[c, "CTRL"] <- NA
		results_fisher[c, "CTRL_M"] <- NumCtrl_M
		results_fisher[c, "CTRL_F"] <- NA
		results_fisher[c, "OR"] <- NA
		results_fisher[c, "LCI"] <- NA
		results_fisher[c, "UCI"] <- NA
		results_fisher[c, "P"] <- NA

		# Increase counter
		c <- c + 1
	}
}

# Claculate the threshold for significance
thr <- 0.05/(55+1)
print(paste0("There are 55 diseases and one burden assessed in both sexes (main CNV-GWAS), setting the threshold for significance at ", round(thr, 6)))

# Disease with significant sex difference
print(paste0("Diseases with significant sex difference: ", nrow(main_df_fisher[which(main_df_fisher$P <= thr), ])))
print(main_df_fisher[which(main_df_fisher$P <= thr), c(1,2,6,7,9,10,11,14)])



#################################################
### Save data ###################################

fwrite(results_fisher, "/cauwerx/gwas/04_phenotypes/data/prevalence/sex_differences_prevalence_fisher.txt", row.names = F, col.names = T, sep = "\t", quote = F)
