==> 04_phenotypes/03_time_to_event.R <==
# Determine time to diagnosis for each case.

#################################################
### Libraries ###################################
library(data.table)
library(dplyr)
library(tidyr)


#################################################
### STEP 1: Set parameters ######################

# Trailing argument
args <- commandArgs(trailingOnly = T)
p <- as.numeric(args[1])
print(paste0("Analysis #", p))

# Date of last diagnsis across all diagnoses (field 41280 (all ICD-10 codes); downladed: 05/2022)
date_last_measure <- "2021-09-30"

# List all phenotypes to analyze; Can be generated from Table S1 (column 1)
pheno_list <- as.data.frame(fread("/data/phenotypes/ICD10_binary_traits_definition.txt"))$phenotype

# Determine the disease to analyze
p <- pheno_list[p]
rm(pheno_list)

# Specify the analyzed sex depending on the disease
sex <- "All"
if (p %in% "PC") {sex <- "M"}
if (p %in% c("BC", "OC", "menstruation", "endometriosis")) {sex <- "F"}
print(paste0("Analyzing: ", p, " (in ", sex, ")"))



#################################################
### STEP 2: Load general data ###################

# Load disease data and select relevant column as the indiactor (case vs ctrl); File generated by "04_phenotypes/01_ICD10_extraction.R"
pheno <- as.data.frame(fread(paste0("/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_", sex,".txt")))
pheno <- pheno[, names(pheno) %in% c("IID", p)]
colnames(pheno)[2] <- "INDICATOR"
pheno$INDICATOR <- factor(pheno$INDICATOR, levels = c(1,2), labels = c(0,1))
pheno <- pheno[pheno$INDICATOR %in% c("0", "1"), ]

# Load disease definition file; Can be generated from Table S1
pheno_definition <- as.data.frame(fread("/data/phenotypes/ICD10_binary_traits_definition.txt", header = T, select = c(1,2), col.names = c("PHENO", "CODES")))
codes_pheno <- gsub("\\.", "", unlist(strsplit(pheno_definition[which(pheno_definition$PHENO == p), "CODES"], split = ", ")))
rm(pheno_definition)

# Define age at birth based on month/year of birth (birthday assumed to be the 15th); "ukb21067.csv" contains phenotypic data month (field 52) and year (field 34) of birth available from the UKBB portal
header <- as.data.frame(fread("/data/UKBB/pheno/ukb21067.csv", header = T, nrow = 0))
col <- grep("^34-|^52", names(header))
age_birth <- as.data.frame(fread("/data/UKBB/pheno/ukb21067.csv", header = T, select = c(1, col), col.names = c("IID", "Year", "Month"))) 
age_birth$DateBirth <- paste0(age_birth$Year, "-", age_birth$Month, "-15")
age_birth <- age_birth[, c(1,4)]
rm(header, col)



#################################################
### STEP 3: Extract age at diagnosis ############

# Diagnosis and date at diagnosis data; "ukb52238.csv" contains phenotypic data for field 41270 (all ICD-10 codes) and 41280 (datt at diagnosis) available from the UKBB portal (downloaded: 05/2022)
header <- as.data.frame(fread("/data/UKBB/pheno/ukb52238.csv", header = T, nrow = 0))

# ICD10 diagnosis (#41270)
col_ICD <- grep("^41270-", names(header))
ICD <- as.data.frame(fread("/data/UKBB/pheno/ukb52238.csv", header = T, select = c(1, col_ICD), colClass = "character"))
ICD[ICD==""] <- NA
ICD <- na.omit(gather(ICD, Instance, Code, 2:ncol(ICD)))
ICD <- unique(ICD[, c(1:3)])
ICD$Instance <- gsub("41270-", "", ICD$Instance)

# Age at diagnosis (#41280)
col_ICD_date <- grep("^41280-", names(header))
ICD_date <- as.data.frame(fread("/data/UKBB/pheno/ukb52238.csv", header = T, select = c(1, col_ICD_date), colClass = "character"))
ICD_date[ICD_date==""] <- NA
ICD_date <- na.omit(gather(ICD_date, Instance, Date, 2:ncol(ICD_date)))
ICD_date <- unique(ICD_date[, c(1:3)])
ICD_date$Instance <- gsub("41280-", "", ICD_date$Instance)

# Match and merge ICD10 codes and age at diagnosis
age_at_diagnosis <- na.omit(left_join(ICD, ICD_date, by = c("eid", "Instance"))) 
rm(header, col_ICD, col_ICD_date, ICD, ICD_date)



#################################################
### STEP 4: Determine ALHM ######################
# ALHM: "age at last healthy measurement"

# Part I: ALHM for samples WITHOUT diagnosis
pheno <- left_join(pheno, age_birth, by = "IID") 
pheno$ALHM <- as.numeric(as.Date(date_last_measure) - as.Date(pheno$DateBirth))/365.25

# Part II: ALHM for samples WITH diagnosis; Earliest date at first diagnosis for any diagnosis on the inclusion list received by the individual
for(eid in pheno[which(pheno$INDICATOR == "1"), "IID"]) {

	# Select the earliest date at which diagnosis was received
	df_eid <- age_at_diagnosis[which(age_at_diagnosis$eid == eid), ]
	date_diagnosis <- min(as.Date(df_eid[grep(paste0("^", codes_pheno, collapse = "|"), df_eid$Code), "Date"]))

	# Calculate the age at first diagnosis and set to missing those with no report date
	if(date_diagnosis == Inf) {
		pheno[which(pheno$IID == eid), "ALHM"] <- NA
	} else {
		pheno[which(pheno$IID == eid), "ALHM"] <- as.numeric(as.Date(date_diagnosis) - as.Date(pheno[which(pheno$IID == eid), "DateBirth"]))/365.25
	}
}
rm(df_eid, date_diagnosis)

# Clean the dataframe
pheno <- na.omit(pheno[, c(1,2,4)])



#################################################
### Save data ###################################

fwrite(pheno, paste0("/cauwerx/gwas/04_phenotypes/data/time_to_event/time_to_event_", sex, ".", p, ".txt.gz"), col.names = T, row.names = F, quote = F, sep = "\t")
