==> 04_phenotypes/04_summary.R <==
# Summarize the number of cases and average age at onset for the studied diseases

#################################################
### Libraries ###################################
library(data.table)
library(dplyr)
library(tidyr)
library(stringr)
library(tibble)



#################################################
### STEP 1: Create an empty dataframe ###########

# Create an empty dataframe to store results
df <- data.frame(PHENO = NA, CASES = NA, CTRL = NA, MISSING = NA, MEAN_onset = NA, SD_onset = NA, MEDIAN_onset = NA, COUNT_onset = NA)
age_onset <- data.frame()



#################################################
### STEP 2: Case/Ctrl/Missing counts ############

# Load disease diagnoses; File generated by "04_phenotypes/01_ICD10_extraction.R"
cases <- as.data.frame(fread("/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_All.txt", header = T, select = c(1:57)))
cases_M <- as.data.frame(fread("/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_M.txt", header = T, select = c(1,4)))
cases <- left_join(cases, cases_M, by = "IID"); rm(cases_M)
cases_F <- as.data.frame(fread("/cauwerx/gwas/04_phenotypes/data/pheno_ICD10_F.txt", header = T, select = c(1:3,60:61)))
cases <- left_join(cases, cases_F, by = "IID"); rm(cases_F)

# Loop over each disorder
for (i in 1:(ncol(cases)-1)) {

	# Define the disease
	p <- names(cases)[i+1]
	print(paste0("Analyzing: ", p))

	# Count the number of occurances
	t <- table(cases[, p], useNA = "always")
	print(t)

	# Assign values to summary dataframe
	df[i, "PHENO"] <- p
	df[i, "CASES"] <- t[which(names(t) == 2)]
	df[i, "CTRL"] <- t[which(names(t) == 1)]
	df[i, "MISSING"] <- t[which(is.na(names(t)))]

}; rm(i, p, t)



#################################################
### STEP 3: Age at onset information ############

# Loop over each disorder
for (i in 1:nrow(df)) {

	# Define the phenotype & sex
	p <- df[i, "PHENO"]
	sex <- "All"; if(p %in% c("OC", "BC", "menstruation", "endometriosis")) {sex <- "F"}; if(p %in% c("PC")) {sex <- "M"}
	print(paste0("Analyzing: ", p))

	# Load age at onset data; File generated by "04_phenotypes/03_time_to_event.R"
	onset <- as.data.frame(fread(paste0("/cauwerx/gwas/04_phenotypes/data/final/time_to_event/time_to_event_", sex, ".", p, ".txt.gz")))

	# Subset cases
	onset <- onset[which(onset$INDICATOR == 1), ]
	print(paste0("Number of cases: ", nrow(onset)))

	# Assign values to summary dataframe
	df[i, "MEAN_onset"] <- mean(onset$ALHM)
	df[i, "SD_onset"] <- sd(onset$ALHM)
	df[i, "MEDIAN_onset"] <- median(onset$ALHM)
	df[i, "COUNT_onset"] <- nrow(onset)

	# Save the age at onsets
	onset$PHENO <- p
	age_onset <- rbind(age_onset, onset[, c(1,3,4)])

}; rm(p, sex, onset)


#################################################
### Save data ###################################

fwrite(df, "/cauwerx/gwas/04_phenotypes/data/summary_cases_age_at_onset.txt", col.names = T, row.names = F, quote = F, sep = "\t")
fwrite(age_onset, "/cauwerx/gwas/04_phenotypes/data/summary_ages_at_onset.txt", col.names = T, row.names = F, quote = F, sep = "\t")
